SimpleExpressions.jl

SimpleExpressions.SimpleExpressions — Module

SimpleExpressions

A very lightweight means to create callable functions using expressions.

The @symbolic macro, the lone export, can create a symbolic variable and optional symbolic parameter. When expressions are created with these variables, evaluation is deferred until the expression is called like a function. The expressions subtype Function so are intended to be useful with Julia's higher-order functions.

The expressions can be called as a univariate function, u(x), a univariate function with parameters, u(x, p), or as a bivariate function, u(x,y) (with y being a parameter). These are all typical calling patterns when a function is passed to a numeric routine. These calls substitute in for the symbolic value (and parameter) when not specified as nothing. (To substitute in for just the parameter, either u(nothing, value) or u(:,value).)

There are no performance claims, this package is all about convenience. Similar convenience is available in some form with SymPy, SymEngine, Symbolics, etc. As well, placeholder syntax is available in Underscores.jl, Chain.jl, DataPipes.jl etc., This package only has value in that it is very lightweight and, hopefully, intuitively simple.

An extension is provided for functions in SpecialFunctions.

An extension is provided for TermInterface which should allow the use of Metatheory to rewrite terms.

Example

using SimpleExpressions
@symbolic x
map(x^2, (1, 2)) # (1, 4)

using Plots
@symbolic x p
u = x^5 - x - p
plot(u(:, 1), 0, 1.5) # substitute in for p
plot!(u(:, 2))

SimpleExpressions.D — Method

D(::AbstractSymbolic)

Finds derivative of a symbolic expression.

assumes a symbolic value is a scalar and takes derivative with respect to that; symbolic parameters are assumed to be constants

There is no simplification, so the output is not necessarily friendly

limited to a select set of functions

Example

julia> @symbolic x p
(x, p)

julia> D(exp(sin(x)))
(1 * cos(x)) * exp(sin(x))

julia> D(D(sin(x))) + sin(x) # no simplification!
(-(sin(x))) + sin(x)

```

SimpleExpressions.@symbolic — Macro

@symbolic x [p]

Create a symbolic variable and optional symbolic parameter.

Expressions and equations

Expressions created using these variables subclass Function so may be used where functions are expected.

The ~ infix operator can be used to create equations, which, by default, are treated as lhs - rhs when used as functions.

Extended help

Example

using SimpleExpressions
@symbolics x
u = x^5 - x - 1
u(2) # 29 call is u(x)

@symbolic x p
u = sum(x .* p)
u(2, [1,2]) # 6  call is u(x, p)

Calling with nothing in a slot leaves the variable

@symbolic x p
u = cos(x) - p*x
u(nothing, 2)  # cos(x) - 2 * x
u(:, 2)        #  cos(x) - 2 * x, alternate calling form
u(pi, nothing) # -1.0 - p * π

The main use is as an easier-to-type replacement for anonymous functions, though with differences:

1 |> sin(x) |> x^2  # sin(1)^2
u = cos(x) - p*x
2 |> u(:, 3) # u(2,3) alternative

map(x^2, (1, 2)) # (1,4)

Can be used with other packages, to simplify some function calls at the expense of being non-idiomatic:

using Roots
@symbolic x p
find_zero(x^5 - x - 1, 1)     # 1.167...
find_zero(x^5 - x ~ p, 1, 4)  # 1.401...

using ForwardDiff
Base.adjoint(𝑓::Function) = x -> ForwardDiff.derivative(𝑓, x)
u = x^5 - x - 1
find_zero((u,u'), 1, Roots.Newton()) # 1.167...

Or

using Plots
plot(x^5 - x - 1, 0, 1.5)

Or using both positions, so that we call as a bivariate function:

@symbolic x y
xs = ys = range(-5, 5, length=100)
contour(xs, ys, x^2 - y^2 + 2x*y)

Symbolic derivatives can be taken with respect to the symbolic value, symbolic parameters are treated as constant.

@symbolic x p
D = SimpleExpressions.D  # not exported
u = x^5 - p*x - 1
D(u)           # (5 * (x ^ 4)) - p
u = u(:, 1)    # set parameter
a, b = 1, 2
find_zeros(D(u) ~ (u(b)-u(a)) / (b-a), (a,b)) # [1.577…]

Extended help

Using this is a convenience for simple cases. It is easy to run into idiosyncrasies.

Expressions are not functions in terms of scope

Unlike functions, expressions are defined with variables at the time of definition, not when called. For example, with a clean environment:

@symbolic x
u = m*x + b    # errors, `m` not defined
f(x) = m*x + b # ok
m, b = 1, 2
u = m*x + b    # defined using `m` amd `b` at time of assignment
u(3)           # 1 * 3 + 2
f(3)           # 1 * 3 + 2 values of `m` and `b` when called
m, b = 3, 4
u(3)           # still computing 1 * 3 + 2
f(3)           # computing 3 * 3 + 4, using values of `m` and `b` when called

Symbolic values are really singletons

Though one can make different symbolic variables, they are all indistinguishable for purposes of evaluation:

@symbolic x
@symbolic y    # both x, y are `Symbolic` type
u = x + 2y
u(3)           # 9 or 3 + 2*3

Similarly for symbolic parameters. The variables may be used as containers though, e.g. u=sum(xi*pi for (xi, pi) in zip(x,p)); u((1,2),(3,4)).

Broadcasting as a function

There is a difference – which needs to be corrected – where it is best to wrap the expression in a container for broadcasting. We can see it here in this artificial example:

@symbolic x
map(x^2, [1,2])    # [1, 4]
map.(x^2, [1,2])   # map.(x^2, [1, 2]) ... not desirable
map.([x^2], [1,2]) # [1, 4]

SimpleExpressions.@symbolic_expression — Macro

@symbolic_expression expr

Take a function call and return a symbolic (delayed) expression.

Example

@symbolic x
u = @symbolic_expression quadgk(sin, 0, x)

# from ?foldl
u = @symbolic_expression foldl(=>, @symbolic_expression(1:x))
u(4) # ((1 => 2) => 3) => 4

Not exported.